Starting the modulariszation of providers
* Split out providers into separate modules
* Removed much cruft in utils.providers
