+++ cfme/tests/infrastructure/test_providers.py
@@ -7,10 +7,10 @@ import pytest
 import cfme.web_ui.flash as flash
 import utils.error as error
 from cfme.exceptions import FlashMessageException
-from cfme.infrastructure.provider import (discover, Provider, VMwareProvider, RHEVMProvider,
-    wait_for_a_provider)
+from cfme.infrastructure.provider import discover, wait_for_a_provider, Provider
+from cfme.infrastructure.provider.rhevm import RHEVMProvider
+from cfme.infrastructure.provider.virtualcenter import VMwareProvider
 from utils import testgen, providers, version
-from utils.providers import get_credentials_from_config
 from utils.update import update
 
 
@@ -185,7 +185,11 @@ def test_provider_add_with_bad_credentials(provider):
     Metadata:
         test_flag: crud
     """
-    provider.credentials['default'] = get_credentials_from_config('bad_credentials')
+    provider.credentials['default'] = provider.Credential(
+        principal='bad',
+        secret='reallybad',
+        verify_secret='reallybad'
+    )
     if isinstance(provider, VMwareProvider):
         with error.expected('Cannot complete login due to an incorrect user name or password.'):
             provider.create(validate_credentials=True)
